requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==5.0.0
python-dotenv==1.0.1
pandas==2.2.1
openpyxl==3.1.2
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests-oauthlib==1.3.1 